RepositoryAPI/requirements.txt
arabic-reshaper==2.1.3
asgiref==3.4.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9
cryptography==36.0.1
Django==3.2.8
django-cors-headers==3.11.0
django-crispy-forms==1.13.0
django-environ==0.8.1
django-ranged-response==0.2.0
django-simple-captcha==0.5.14
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
environ==1.0
future==0.18.2
html5lib==1.1
idna==3.3
midtransclient==1.4.0
mysqlclient @ file:///C:/Dev/Django/mysqlclient-1.4.6-cp39-cp39-win_amd64.whl
Pillow==8.4.0
pycparser==2.21
PyJWT==1.5.3
PyPDF2==1.26.0
python-bidi==0.4.2
pytz==2021.3
reportlab==3.6.3
requests==2.26.0
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.7
webencodings==0.5.1
xhtml2pdf==0.2.5
